friendrequests page updated
made the input to store the friends name, then when pressed to send the request, it fetches the id and finds the username, then its add to the recipient friends request list. created a method to get usernamebyId in databasemethods, with its getfriendrequests and sendfriendrequest. accepting or rejecting the friend request is left. search for friends is TO DO list.